feat: delete ClaudeAgent mock implementation and add anti-mock hook
BREAKING CHANGE: Removed ClaudeAgent entirely - was generating fake sub-queries and wrapping real drone results in boilerplate

Changes:
- Delete claude_agent.go - 251 lines of mock code removed
- Update orchestrator to send research topic directly to drones
- Drones now handle their own query formulation and research methods
- Replace fake report generation with real data aggregation from drone results
- Add pre-commit hook to block mock/placeholder code from being committed
- Update tests to remove ClaudeAgent dependencies

Before: User → ClaudeAgent (fake queries) → Drones (real work) → ClaudeAgent (boilerplate wrap) → User
After: User → Drones (real work) → Queue → User (real results)

The queue already collects real research results from drones. No need for a fake AI layer.
