--- cmd/widescreen-research-mcp/orchestrator/claude_agent.go
@@ -1,251 +1,0 @@
-// Copyright (c) 2025 glassBead-tc and contributors
-// SPDX-License-Identifier: MIT
-
-package orchestrator
-
-import (
-	"context"
-	"fmt"
-	"log"
-	"time"
-
-	"github.com/glassBead-tc/widescreen-research/cmd/widescreen-research-mcp/schemas"
-)
-
-// ClaudeAgent manages AI-powered orchestration using Claude
-type ClaudeAgent struct {
-	// In a real implementation, this would use the Claude SDK
-	// For now, we'll create a mock implementation
-	apiKey string
-}
-
-// NewClaudeAgent creates a new Claude agent
-func NewClaudeAgent() *ClaudeAgent {
-	return &ClaudeAgent{
-		apiKey: getEnvOrDefault("CLAUDE_API_KEY", ""),
-	}
-}
-
-// Initialize initializes the Claude agent
-func (a *ClaudeAgent) Initialize(ctx context.Context) error {
-	if a.apiKey == "" {
-		log.Println("Warning: CLAUDE_API_KEY not set, using mock Claude agent")
-	}
-	return nil
-}
-
-// GenerateSubQueries uses the AI to break a high-level topic into specific sub-queries.
-func (a *ClaudeAgent) GenerateSubQueries(ctx context.Context, topic string, numQueries int) ([]string, error) {
-	// In a real implementation, this would use Claude. For now, mock data.
-	log.Printf("Generating %d mock sub-queries for topic: %s", numQueries, topic)
-	if topic == "Top 3 AI Companies" {
-		return []string{
-			"Detailed analysis of OpenAI's business model, products, and recent controversies.",
-			"Financial performance and strategic initiatives of Google's AI division (DeepMind, Google AI).",
-			"Overview of Microsoft's AI strategy, focusing on its partnership with OpenAI and Azure AI services.",
-		}, nil
-	}
-
-	// Default mock data
-	var queries []string
-	for i := 1; i <= numQueries; i++ {
-		queries = append(queries, fmt.Sprintf("Sub-query %d for %s", i, topic))
-	}
-	return queries, nil
-}
-
-// GenerateReport generates a research report from collected data
-func (a *ClaudeAgent) GenerateReport(ctx context.Context, config *schemas.ResearchConfig, results []schemas.DroneResult, analysis *DataAnalysis) (*schemas.ResearchReport, error) {
-	// Process results into a structured report
-
-	report := &schemas.ResearchReport{
-		Title:       fmt.Sprintf("Research Report: %s", config.Topic),
-		Executive:   a.generateExecutiveSummary(config, results, analysis),
-		Sections:    a.generateReportSections(config, results, analysis),
-		Methodology: a.generateMethodologySection(config),
-		Data:        a.aggregateData(results),
-		Metadata: schemas.ReportMetadata{
-			ResearchTopic:   config.Topic,
-			ResearcherCount: config.ResearcherCount,
-			Duration:        analysis.Duration,
-			DataPoints:      len(results),
-			Sources:         a.extractSources(results),
-			Metrics:         analysis.Metrics,
-		},
-	}
-
-	return report, nil
-}
-
-// generateExecutiveSummary generates an executive summary
-func (a *ClaudeAgent) generateExecutiveSummary(config *schemas.ResearchConfig, results []schemas.DroneResult, analysis *DataAnalysis) string {
-	summary := fmt.Sprintf("Executive Summary: %s\n\n", config.Topic)
-	summary += fmt.Sprintf("This research was conducted using %d parallel research drones over %v.\n\n",
-		config.ResearcherCount, analysis.Duration)
-
-	summary += "Key Findings:\n"
-	for i, insight := range analysis.TopInsights {
-		if i >= 3 {
-			break
-		}
-		summary += fmt.Sprintf("- %s\n", insight)
-	}
-
-	return summary
-}
-
-// generateReportSections generates report sections
-func (a *ClaudeAgent) generateReportSections(config *schemas.ResearchConfig, results []schemas.DroneResult, analysis *DataAnalysis) []schemas.ReportSection {
-	sections := []schemas.ReportSection{
-		{
-			Title:   "Introduction",
-			Content: a.generateIntroduction(config),
-		},
-		{
-			Title:    "Key Findings",
-			Content:  a.generateKeyFindings(results, analysis),
-			Insights: analysis.TopInsights,
-		},
-		{
-			Title:   "Data Analysis",
-			Content: a.generateDataAnalysis(analysis),
-			Data:    analysis.Statistics,
-		},
-		{
-			Title:   "Conclusions",
-			Content: a.generateConclusions(config, analysis),
-		},
-	}
-
-	return sections
-}
-
-// Helper methods for report generation
-
-func (a *ClaudeAgent) generateIntroduction(config *schemas.ResearchConfig) string {
-	return fmt.Sprintf("This report presents the findings from a comprehensive research study on '%s'. "+
-		"The research was conducted using %d parallel research agents with a %s depth approach.",
-		config.Topic, config.ResearcherCount, config.ResearchDepth)
-}
-
-func (a *ClaudeAgent) generateKeyFindings(results []schemas.DroneResult, analysis *DataAnalysis) string {
-	findings := "Based on the analysis of data from all research drones, the following key findings emerged:\n\n"
-
-	// Group findings by status
-	successCount := 0
-	for _, result := range results {
-		if result.Status == "completed" {
-			successCount++
-		}
-	}
-
-	findings += fmt.Sprintf("- Successfully collected data from %d out of %d drones\n", successCount, len(results))
-	findings += fmt.Sprintf("- Identified %d key patterns across the dataset\n", len(analysis.Patterns))
-
-	return findings
-}
-
-func (a *ClaudeAgent) generateDataAnalysis(analysis *DataAnalysis) string {
-	return fmt.Sprintf("The data analysis revealed %d patterns with an average confidence of %.2f. "+
-		"Statistical analysis shows %v unique data points collected.",
-		len(analysis.Patterns), analysis.AverageConfidence, analysis.Statistics["total_data_points"])
-}
-
-func (a *ClaudeAgent) generateConclusions(config *schemas.ResearchConfig, analysis *DataAnalysis) string {
-	return fmt.Sprintf("The research on '%s' has provided comprehensive insights through parallel processing. "+
-		"The %s-depth analysis approach yielded %d actionable insights with high confidence.",
-		config.Topic, config.ResearchDepth, len(analysis.TopInsights))
-}
-
-func (a *ClaudeAgent) generateMethodologySection(config *schemas.ResearchConfig) string {
-	return fmt.Sprintf("This research employed a distributed approach using %d parallel research drones. "+
-		"Each drone was tasked with specific aspects of the research topic '%s'. "+
-		"The %s-depth methodology ensured comprehensive coverage while maintaining efficiency.",
-		config.ResearcherCount, config.Topic, config.ResearchDepth)
-}
-
-func (a *ClaudeAgent) aggregateData(results []schemas.DroneResult) map[string]interface{} {
-	aggregated := make(map[string]interface{})
-
-	// Collect all data from successful drones
-	var allData []map[string]interface{}
-	for _, result := range results {
-		if result.Status == "completed" && result.Data != nil {
-			allData = append(allData, result.Data)
-		}
-	}
-
-	aggregated["drone_data"] = allData
-	aggregated["total_results"] = len(results)
-	aggregated["successful_results"] = len(allData)
-
-	return aggregated
-}
-
-func (a *ClaudeAgent) extractSources(results []schemas.DroneResult) []string {
-	sourceMap := make(map[string]bool)
-
-	for _, result := range results {
-		if sources, ok := result.Data["sources"].([]interface{}); ok {
-			for _, source := range sources {
-				if s, ok := source.(string); ok {
-					sourceMap[s] = true
-				}
-			}
-		}
-	}
-
-	sources := make([]string, 0, len(sourceMap))
-	for source := range sourceMap {
-		sources = append(sources, source)
-	}
-
-	return sources
-}
-
-// AnalyzeSequentialThinking performs sequential thinking analysis
-func (a *ClaudeAgent) AnalyzeSequentialThinking(ctx context.Context, problem string, context string) (*schemas.SequentialThinkingResponse, error) {
-	// Mock implementation of sequential thinking
-	thoughts := []schemas.ThoughtStep{
-		{
-			Step:       1,
-			Thought:    "Understanding the problem: " + problem,
-			Reasoning:  "First, we need to clearly understand what we're trying to solve",
-			Confidence: 0.95,
-		},
-		{
-			Step:       2,
-			Thought:    "Analyzing the context and constraints",
-			Reasoning:  "Context provides important boundaries and requirements",
-			Confidence: 0.90,
-		},
-		{
-			Step:       3,
-			Thought:    "Generating potential solutions",
-			Reasoning:  "Based on the problem and context, we can identify approaches",
-			Confidence: 0.85,
-		},
-	}
-
-	return &schemas.SequentialThinkingResponse{
-		Thoughts:   thoughts,
-		Solution:   "Based on sequential analysis, the recommended approach is to proceed with distributed research",
-		Confidence: 0.88,
-	}, nil
-}
-
-// Shutdown shuts down the Claude agent
-func (a *ClaudeAgent) Shutdown() {
-	// Clean up any resources
-}
-
-// Supporting types
-
-type DataAnalysis struct {
-	Patterns          []schemas.Pattern
-	TopInsights       []string
-	Statistics        map[string]interface{}
-	Duration          time.Duration
-	AverageConfidence float64
-	Metrics           schemas.ResearchMetrics
-}
